The Switch XNA -> FNA
switched out XNA for FNA
changed the spritebatch so the graphics work as intended
changed code in resolution as it was windows forms
added FixLoadAssembly to Modloader to remap loaded mods. added all of the xnaToFna code base, and some other junk for compatibility. linux support not yet added need to work on some stuff there first and its late.