Reworked the way the rebuilt client attributes work
Previously each attribute had to do it's own search for every type and every member in that type for the executing assembly. And this would multiply with the amount of attributes there were. Pretty wasteful, right? Now there's a new system in place which will simply do a single search, then for all the requested attribute searches there will be a dictionary containing the members using it. That way all you have to do is to lookup your attribute in that dictionary and you get a list of every member using it. The actual system works a bit differently than this, but that's basically the gist of how it's a lot better than it used to be.