Added support for properties for AutoConfig
You can now use the attribute on properties, and their getters and setters will be invoked whenever necessary. You can also leave out either if you dont want anything to happen when the value is being loaded or saved, however lacking a getter will cause the value to be saved with the value `null` on the config file.