Added the ProgressValue struct and EasingHelper.cs
The ProgressValue is a struct designed to make it easy to deal with values that have a maximum and a minimum value that they cannot exceed, like progress bar values or animation states.

EasingHelper.cs provides a library of easing functions that can be used for duck game. This will make creating coded animations or highlights a lot simpler, since now all you need is to just call a function from the class and you now have the eased value.

Also added `double` clamping in Maths.cs