Added DGCL lexer and syntax highlighter
This commit includes the DebugTablet.cs which is a sort of text editor to debug how DGCL would look. This would be removed later and probably replaced with a toggle in the DevConsole itself.

This also includes a system for locking duck inputs (doesn't include controller inputs) on demand. Meaning you can add your movement locking service to LockMovementQueue and your duck wont move, and remove it when needed. This can be useful for creating prompts that require keyboard input without moving the duck.