Regain UI classes' compatibility with mods
Thank Dan for pointing these out:

I apologize for disregarding the modding compatibility when touching the UI classes before - It did improve the code but it also changed its logic, meaning that if a mod uses vanilla UI code it would have problem working with DGR. So the first thing I had to do was to revert basically all changes in that commit where the algorithm for some text positioning was rewritten (wherefore the compatibility with mods was ruined).

But I still see the potential worth to keep the fixed version so that in the future we'll be able to build stuff using it without struggling with the buggy vanilla UI framework. To do this, as Dan suggested, I added some *new* classes (the LUI ones) that save the feature I want to keep. In this way, the vanilla UI part is actually maintained and untouched, and thus compatible with mods, while the new feature is still available for futural uses.