INFINITE FPS
This isn't ideal but all things considered it covers a large part of the game and is good enough for fulltime use. There is some performance overhead with storing and updating the lerp values on each sprite/particle but it's small I hope. (probably worth optimizing for the particles)

Known issues:
-Things drawn with Textures (aka a tex2d+ position draw call) will not be interpolated, will need to be added manually.
-Things that get teleported will be interpolated (sometimes this is nice)
-Graphics Culling will cause sprites to pop rarely
-Curved bullets are missing
-Some specific particles are missing
-Lots of UI menus are missing, only a few work
-There might be somewhere with fucked drawing that I didn't find during my testing which has to be excluded from interpolation